Harden queue detail lookups and docs

+++ crates/sr-common/src/api/queue_job.rs
@@ -137,7 +137,7 @@ pub struct MatchResultRow {
 #[derive(Debug, Clone, Serialize)]
 pub struct InteractionLogRow {
     pub id: i64,
-    pub match_result_id: Option<i64>,
+    pub match_result_id: Option<i32>,
     pub talent_id: i64,
     pub project_id: i64,
     pub match_run_id: Option<String>,
@@ -154,7 +154,7 @@ pub struct InteractionLogRow {
 pub struct FeedbackEventRow {
     pub id: i64,
     pub interaction_id: Option<i64>,
-    pub match_result_id: Option<i64>,
+    pub match_result_id: Option<i32>,
     pub match_run_id: Option<String>,
     pub engine_version: Option<String>,
     pub config_version: Option<String>,
